[Bronze V] Title: 이상한 기호, Time: 212 ms, Memory: 17664 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner kb =  new Scanner(System.in);
 
-        int num1 = kb.nextInt();
-        int num2 = kb.nextInt();
+        Long num1 = kb.nextLong();
+        Long num2 = kb.nextLong();
 
         System.out.println((num1 + num2) * (num1 - num2));
     }
